Makes ion-c-sys an opt-in feature
`ion-c-sys` provides bindings to the `ion-c` library, allowing Rust
applications to use `ion-c`'s readers and writers to work with Ion data.
Now that the native Rust readers and writers are passing the `ion-tests`
conformance suite, `ion-c-sys` is no longer mandatory.

This PR introduces a new feature, "ion-c-sys", that allows users to opt
into the functionality and dependencies that come with `ion-c-sys`.

For users that do not require `ion-c-sys`, this eliminates the need to
install a C toolchain on their development machine (cmake, clang, etc).
It also cuts build times and test runtimes by more than half.
