+++ ion-hash/Cargo.toml
@@ -20,7 +20,7 @@ version = "0.0.7"
 edition = "2021"
 
 [dependencies]
-ion-rs = { path = "../", version = "0.10" }
+ion-rs = { path = "../", version = "0.10", features = ["ion_c"]}
 ion-c-sys = { path = "../ion-c-sys", version = "0.4" }
 num-bigint = "0.3"
 digest = "0.9"
